requirements.txt
streamlit
google-generativeai
bcrypt
pandas
sqlalchemy
fpdf
